+++ src/includes/classes/SCore/Utils/StylesScripts.php
@@ -490,15 +490,15 @@ class StylesScripts extends Classes\SCore\Base\Core
         $data = [
             'styles' => [
                 'semantic-ui' => [
-                    'version' => '2.2.7',
-                    'url'     => '//cdnjs.cloudflare.com/ajax/libs/semantic-ui/%1$s/semantic.min.css',
+                    'ver' => $this->cv,
+                    'url' => $this->c::appWsCoreUrl('/client-s/semantic/dist/semantic.min.css'),
                 ],
             ],
             'scripts' => [
                 'semantic-ui' => [
-                    'version' => '2.2.7',
-                    'deps'    => ['jquery'],
-                    'url'     => '//cdnjs.cloudflare.com/ajax/libs/semantic-ui/%1$s/semantic.min.js',
+                    'ver'  => $this->cv,
+                    'deps' => ['jquery'],
+                    'url'  => $this->c::appWsCoreUrl('/client-s/semantic/dist/semantic.min.js'),
                 ],
             ],
         ];