+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## $v
 
+- Adding `s::enqueueStripeLibs()`.
+- Adding `s::enqueueNoUiSliderLibs()`.
 - Updating to latest WebSharks Core, which now includes Stripe integration.
 
 ## v170311.43193