+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v170329.20871
 
 - Adding `s::enqueueStripeLibs()`.
 - Adding `s::enqueueNoUiSliderLibs()`.